float 2 double

+++ assignment5/src/diffusion.c
@@ -1,6 +1,6 @@
 #include "diffusion.h"
 
-void diffusion_step(float* prev_step, float* next_step, int rows, int cols, float c) {
+void diffusion_step(double* prev_step, double* next_step, int rows, int cols, double c) {
     for (size_t i = 1; i < rows; ++i) {
         for (size_t j = 1; j < cols; ++j) {
             int index = i * (cols + 2) + j;
