+++ assignment5/src/read_input.h
@@ -5,6 +5,6 @@
 #include <stdlib.h>
 
 // Returns padded matrix, size of box
-float * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols);
+double * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols);
 
 #endif
